+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Highlights are marked with a pancake 🥞
 - node: Cancel processed stream tasks on drop of stream handles [#1300](https://github.com/p2panda/p2panda/pull/1300)
 - spaces: Support partial merge of auth state into a space [#1299](https://github.com/p2panda/p2panda/pull/1299)
 - node: Allow graceful closure of sync sessions [#1307](https://github.com/p2panda/p2panda/pull/1307)
+- node: API for promoting and demoting space members [#1311](https://github.com/p2panda/p2panda/pull/1311)
 
 ### Changed